pyproject.toml
[tool.poetry]
name = "qq-bot"
version = "0.0.1"
description = ""
authors = ["nullptr <wfx.weifanxiang@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
graia-application-mirai = "^0.19.0"
graia-saya = "^0.0.9"
async-google-trans-new = "^1.4.1"
pyyaml = "^5.4.1"
graia-scheduler = "^0.0.4"

[tool.poetry.dev-dependencies]
yapf = "^0.31.0"
pylint = "^2.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
default = true
url = "https://mirrors.aliyun.com/pypi/simple"
